refactor(navigation): updated interfaces to handle navigation and fixed tsx types errors in app
