+++ package.json
@@ -17,6 +17,7 @@
   },
   "dependencies": {
     "@react-navigation/native": "^6.0.10",
+    "@react-navigation/native-stack": "^6.9.12",
     "@react-navigation/stack": "^6.2.1",
     "@types/react-native-dotenv": "^0.2.0",
     "axios": "^0.27.2",
