+++ src/interfaces/movies.ts
@@ -1,6 +1,3 @@
-import {StackScreenProps} from '@react-navigation/stack';
-import {RootStackParams} from 'interfaces';
-
 /* API types */
 /* Now Playing */
 export interface MovieDBResponseAPI {
@@ -51,9 +48,6 @@ export interface MoviesCarrouselProps {
   movies: Movie[];
   title?: string;
 }
-
-export interface IDetailScreenProps
-  extends StackScreenProps<RootStackParams, 'DetailScreen'> {}
 
 // Generated by https://quicktype.io
