+++ src/interfaces/navigation.ts
@@ -1,6 +1,12 @@
 import {Movie} from './movies';
+import type {NativeStackScreenProps} from '@react-navigation/native-stack';
 
 export type RootStackParams = {
   HomeScreen: undefined;
-  DetailScreen: Movie;
+  DetailScreen: {movie: Movie};
 };
+
+export type IDetailScreenProps = NativeStackScreenProps<
+  RootStackParams,
+  'DetailScreen'
+>;
