+++ src/navigation/Navigation.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {createStackNavigator} from '@react-navigation/stack';
 import {DetailScreen, HomeScreen} from '@screens/index';
-import {RootStackParams} from 'interfaces';
+import {RootStackParams} from '@interfaces/index';
 
 const Stack = createStackNavigator<RootStackParams>();
