+++ tsconfig.json
@@ -44,6 +44,7 @@
       "@components/*": ["src/components/*"],
       "@constants/*": ["src/constants/*"],
       "@hooks/*": ["src/hooks/*"],
+      "@context/*": ["src/context/*"],
       "@interfaces/*": ["src/interfaces/*"],
       "@navigation/*": ["src/navigation/*"],
       "@screens/*": ["src/screens/*"],
